requirements.txt
numpy>=1.17.0
pandas>=0.25.0
pvlib>=0.9.0
python-dateutil>=2.8.0
pytz>=2022.1
PyYAML>=5.1.2
